Change printf for long + add some bonus

+++ libft/includes/ft_printf_fd.h
@@ -12,7 +12,7 @@ int		ft_treat_type(int fd, const char c, va_list args);
 int		print_addr_fd(unsigned long long addr, int fd);
 int		print_str_fd(char *s, int fd);
 int		print_nbr_fd(int n, int fd);
-int		print_unbr_fd(unsigned int n, int fd);
+int		print_unbr_fd(unsigned long n, int fd);
 int		print_nbrbase_fd(unsigned long long nbr, int fd, char *base);
 int		ft_is_base_wrong(char *base);
 int		ft_nbrlen_base(unsigned long long nbr, char *base);
